chore(2916): refactor stream handler for phishing warning and legacy
